refactor: add directory structure for main
