+++ src/index.js
@@ -1,7 +1,7 @@
 import { initEventListener } from "./util/initEventListener.js";
 import { HeaderView } from "./view/HeaderView.js";
 import { HistoryView } from "./view/HistoryView.js";
-import { MainView } from "./view/MainView.js";
+import MainView from "./view/Main/index.js";
 
 (function init() {
   const app = document.getElementById("app");
