+++ src/view/Main/Column/CardListView.js
@@ -1,5 +1,5 @@
-import { columnDataTable } from "../../model/model.js";
-import { CardView } from "./CardView.js";
+import { columnDataTable } from "../../../model/model.js";
+import { CardView } from "./Card/CardView.js";
 
 export const CardListView = (columnId) => {
   const { value: cardList } = columnDataTable[columnId];
