+++ src/view/Main/index.js
@@ -1,7 +1,7 @@
-import { columnDataTable } from "../model/model.js";
-import { ColumnView } from "./components/ColumnView.js";
+import { columnDataTable } from "../../model/model.js";
+import ColumnView from "./Column/index.js";
 
-export const MainView = () => {
+const MainView = () => {
   return `
 <main class="main">
     ${Object.keys(columnDataTable)
@@ -10,3 +10,5 @@ export const MainView = () => {
       </main>
 `;
 };
+
+export default MainView;
